# api/openapi/v1/openapi.yaml
openapi: 3.1.0
info:
  title: Project Progress Register API
  version: 0.1.0
  description: >-
    Authoritative contract for the implemented Project Progress Register HTTP API.
    Cookie-authenticated internal API. Login returns a session-bound CSRF token;
    authenticated write requests must send it in X-CSRF-Token.
servers:
  - url: '{basePath}'
    description: Current application origin and deployment prefix.
    variables:
      basePath:
        default: /
        description: Use / for root hosting or the configured BASE_PATH, such as /backend.
...
paths:
  /api/v1/setup/bootstrap:
    post:
      tags: [Identity]
      operationId: bootstrapAdmin
      summary: Create the first Admin exactly once
      description: Available only while no users exist and BOOTSTRAP_TOKEN is configured. The token is never retained.
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/BootstrapRequest'}
            example: {bootstrap_token: replace-with-configured-secret, username: admin, email: admin@example.com, password: correct horse battery staple}
      responses:
        '201':
          description: Initial Admin created.
          content: {application/json: {schema: {$ref: '#/components/schemas/UserResponse'}}}
        '403': {$ref: '#/components/responses/BootstrapDenied'}
        '404': {$ref: '#/components/responses/BootstrapUnavailable'}
        '422': {$ref: '#/components/responses/ValidationFailed'}
  /api/v1/auth/login:
    post:
      tags: [Identity]
      operationId: login
      summary: Start a revocable opaque session
      description: Unknown, disabled, incorrect, and throttled credentials have the same client-visible response. Sets the host-only ppr_session cookie.
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/LoginRequest'}
            example: {identifier: admin, password: correct horse battery staple}
      responses:
        '200':
          description: Authenticated; a new session cookie is set.
          headers:
            Set-Cookie: {schema: {type: string}, description: HttpOnly host-only ppr_session cookie with SameSite=Lax; Secure in production.}
          content: {application/json: {schema: {$ref: '#/components/schemas/SessionResponse'}}}
        '401': {$ref: '#/components/responses/InvalidCredentials'}
        '415': {$ref: '#/components/responses/UnsupportedMediaType'}
  /api/v1/auth/session:
    get:
      tags: [Identity]
      operationId: getCurrentSession
      summary: Recover the authenticated user and CSRF token
      security: [{cookieSession: []}]
      responses:
        '200':
          description: Current non-expired session for an enabled user.
          content: {application/json: {schema: {$ref: '#/components/schemas/SessionResponse'}}}
        '401': {$ref: '#/components/responses/Unauthenticated'}
  /api/v1/auth/logout:
    post:
      tags: [Identity]
      operationId: logout
      summary: Revoke the current session
      description: Requires X-CSRF-Token when a valid session is present. Repeated logout without a session succeeds.
      security: [{cookieSession: []}]
      parameters:
        - in: header
          name: X-CSRF-Token
          required: false
          schema: {type: string}
          description: Required for an authenticated session and obtained from login or session recovery.
      responses:
        '200':
          description: Session is absent or has been revoked and the cookie is cleared.
          content: {application/json: {schema: {$ref: '#/components/schemas/LogoutResponse'}}}
        '403': {$ref: '#/components/responses/CSRFInvalid'}
  /api/v1/auth/password:
    post:
      tags: [Identity]
      operationId: changeOwnPassword
      summary: Replace the authenticated user's password
      description: Revokes every session including the current session and clears must_change_password. The caller must sign in again.
      security: [{cookieSession: []}]
      parameters:
        - {$ref: '#/components/parameters/CSRFToken'}
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/ChangePasswordRequest'}
      responses:
        '200':
          description: Password changed, all sessions revoked, and the cookie cleared.
          content: {application/json: {schema: {$ref: '#/components/schemas/PasswordChangedResponse'}}}
        '401': {$ref: '#/components/responses/Unauthenticated'}
        '403': {$ref: '#/components/responses/CSRFInvalid'}
        '422': {$ref: '#/components/responses/ValidationFailed'}
  /api/v1/admin/users:
    get:
      tags: [User administration]
      operationId: listUsers
      summary: List all internal accounts
      description: Requires an enabled Admin whose temporary password has already been replaced. Results are ordered by username.
      security: [{cookieSession: []}]
      responses:
        '200':
          description: Complete small internal account inventory.
          content: {application/json: {schema: {$ref: '#/components/schemas/UsersResponse'}}}
        '401': {$ref: '#/components/responses/Unauthenticated'}
        '403': {$ref: '#/components/responses/Forbidden'}
    post:
      tags: [User administration]
      operationId: createUser
      summary: Create an Admin or Member
      description: Generates a temporary password returned only in this no-store response. The user must replace it after login.
      security: [{cookieSession: []}]
      parameters:
        - {$ref: '#/components/parameters/CSRFToken'}
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/CreateUserRequest'}
      responses:
        '201':
          description: Account created; copy the temporary password now because it cannot be retrieved.
          headers:
            Cache-Control: {schema: {type: string, const: no-store}}
          content: {application/json: {schema: {$ref: '#/components/schemas/CredentialResponse'}}}
        '401': {$ref: '#/components/responses/Unauthenticated'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '409': {$ref: '#/components/responses/Conflict'}
        '422': {$ref: '#/components/responses/ValidationFailed'}
  /api/v1/admin/users/{user_id}:
    patch:
      tags: [User administration]
      operationId: updateUser
      summary: Change an account role or enabled state
      description: Uses optimistic versioning, revokes all target sessions, and refuses to remove the final enabled Admin.
      security: [{cookieSession: []}]
      parameters:
        - {$ref: '#/components/parameters/UserID'}
        - {$ref: '#/components/parameters/CSRFToken'}
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/UpdateUserRequest'}
      responses:
        '200':
          description: Account updated and target sessions revoked.
          content: {application/json: {schema: {$ref: '#/components/schemas/UserResponse'}}}
        '401': {$ref: '#/components/responses/Unauthenticated'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '404': {$ref: '#/components/responses/NotFound'}
        '409': {$ref: '#/components/responses/Conflict'}
        '422': {$ref: '#/components/responses/ValidationFailed'}
  /api/v1/admin/users/{user_id}/password-reset:
    post:
      tags: [User administration]
      operationId: resetUserPassword
      summary: Generate a temporary password for an account
      description: Replaces the password, sets must_change_password, revokes every target session, and returns the temporary password once.
      security: [{cookieSession: []}]
      parameters:
        - {$ref: '#/components/parameters/UserID'}
        - {$ref: '#/components/parameters/CSRFToken'}
      responses:
        '200':
          description: Password reset; copy the temporary password now because it cannot be retrieved.
          headers:
            Cache-Control: {schema: {type: string, const: no-store}}
          content: {application/json: {schema: {$ref: '#/components/schemas/CredentialResponse'}}}
        '401': {$ref: '#/components/responses/Unauthenticated'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '404': {$ref: '#/components/responses/NotFound'}
  /api/v1/admin/audit/identity:
    get:
      tags: [User administration]
      operationId: listIdentityAudit
      summary: List recent authentication and account audit events
      description: Returns the newest 200 identity, authentication, and account-authorization events. Requires an enabled Admin with no forced password change.
      security: [{cookieSession: []}]
      responses:
        '200':
          description: Newest identity audit events in reverse chronological order.
          content: {application/json: {schema: {$ref: '#/components/schemas/IdentityAuditResponse'}}}
        '401': {$ref: '#/components/responses/Unauthenticated'}
        '403': {$ref: '#/components/responses/Forbidden'}
  /api/v1/projects:
    get:
      tags: [Projects]
      operationId: listProjects
      summary: List projects visible to the current user
      description: Admins receive every project. Members receive only projects with a current membership. Inactive projects remain readable for history.
      security: [{cookieSession: []}]
      responses:
        '200':
          description: Name-ordered authorized project list with current geofence policy when configured.
          content: {application/json: {schema: {$ref: '#/components/schemas/ProjectsResponse'}}}
        '401': {$ref: '#/components/responses/Unauthenticated'}
        '403': {$ref: '#/components/responses/Forbidden'}
    post:
      tags: [Projects]
      operationId: createProject
      summary: Create a project
      description: Admin-only command. Creates an active project without a geofence or members and appends project.created atomically.
      security: [{cookieSession: []}]
      parameters: [{$ref: '#/components/parameters/CSRFToken'}]
      requestBody:
        required: true
        content: {application/json: {schema: {$ref: '#/components/schemas/CreateProjectRequest'}}}
      responses:
        '201':
          description: Project created.
          content: {application/json: {schema: {$ref: '#/components/schemas/ProjectResponse'}}}
        '401': {$ref: '#/components/responses/Unauthenticated'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '422': {$ref: '#/components/responses/ValidationFailed'}
  /api/v1/projects/{project_id}:
    get:
      tags: [Projects]
      operationId: getProject
      summary: Get an authorized project
      description: Members without current membership receive the same 404 as an unknown identifier.
      security: [{cookieSession: []}]
      parameters: [{$ref: '#/components/parameters/ProjectID'}]
      responses:
        '200':
          description: Authorized project and current geofence policy.
          content: {application/json: {schema: {$ref: '#/components/schemas/ProjectResponse'}}}
        '401': {$ref: '#/components/responses/Unauthenticated'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '404': {$ref: '#/components/responses/NotFound'}
    patch:
      tags: [Projects]
      operationId: updateProject
      summary: Replace mutable project fields
      description: Admin-only optimistic update of name, Markdown description, and active state. Projects are deactivated rather than deleted.
      security: [{cookieSession: []}]
      parameters:
        - {$ref: '#/components/parameters/ProjectID'}
        - {$ref: '#/components/parameters/CSRFToken'}
      requestBody:
        required: true
        content: {application/json: {schema: {$ref: '#/components/schemas/UpdateProjectRequest'}}}
      responses:
        '200':
          description: Project updated.
          content: {application/json: {schema: {$ref: '#/components/schemas/ProjectResponse'}}}
        '401': {$ref: '#/components/responses/Unauthenticated'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '404': {$ref: '#/components/responses/NotFound'}
        '409': {$ref: '#/components/responses/Conflict'}
        '422': {$ref: '#/components/responses/ValidationFailed'}
  /api/v1/projects/{project_id}/members:
    get:
      tags: [Projects]
      operationId: listProjectMembers
      summary: List current project Members
      description: Admin-only query. Disabled accounts remain visible so access state is explicit.
      security: [{cookieSession: []}]
      parameters: [{$ref: '#/components/parameters/ProjectID'}]
      responses:
        '200':
          description: Username-ordered current membership list.
          content: {application/json: {schema: {$ref: '#/components/schemas/ProjectMembersResponse'}}}
        '401': {$ref: '#/components/responses/Unauthenticated'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '404': {$ref: '#/components/responses/NotFound'}
  /api/v1/projects/{project_id}/members/{user_id}:
    put:
      tags: [Projects]
      operationId: addProjectMember
      summary: Add a current project Member
      description: Admin-only command. The target must currently be an enabled Member account. Historical membership rows are retained.
      security: [{cookieSession: []}]
      parameters:
        - {$ref: '#/components/parameters/ProjectID'}
        - {$ref: '#/components/parameters/UserID'}
        - {$ref: '#/components/parameters/CSRFToken'}
      responses:
        '201':
          description: Current membership created.
          content: {application/json: {schema: {$ref: '#/components/schemas/ProjectMemberResponse'}}}
        '401': {$ref: '#/components/responses/Unauthenticated'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '404': {$ref: '#/components/responses/NotFound'}
        '409': {$ref: '#/components/responses/Conflict'}
        '422': {$ref: '#/components/responses/ValidationFailed'}
    delete:
      tags: [Projects]
      operationId: removeProjectMember
      summary: End a current project membership
      description: Admin-only command. Closes the temporal membership and immediately removes future Member access without erasing history.
      security: [{cookieSession: []}]
      parameters:
        - {$ref: '#/components/parameters/ProjectID'}
        - {$ref: '#/components/parameters/UserID'}
        - {$ref: '#/components/parameters/CSRFToken'}
      responses:
        '204': {description: Membership ended; no response body.}
        '401': {$ref: '#/components/responses/Unauthenticated'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '404': {$ref: '#/components/responses/NotFound'}
  /api/v1/projects/{project_id}/geofence:
    put:
      tags: [Projects]
      operationId: replaceProjectGeofence
      summary: Create or replace the current site geofence
      description: Admin-only optimistic command. Closes the previous immutable policy version and inserts the next version atomically.
      security: [{cookieSession: []}]
      parameters:
        - {$ref: '#/components/parameters/ProjectID'}
        - {$ref: '#/components/parameters/CSRFToken'}
      requestBody:
        required: true
        content: {application/json: {schema: {$ref: '#/components/schemas/ReplaceGeofenceRequest'}}}
      responses:
        '200':
          description: New current geofence policy.
          content: {application/json: {schema: {$ref: '#/components/schemas/GeofenceResponse'}}}
        '401': {$ref: '#/components/responses/Unauthenticated'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '404': {$ref: '#/components/responses/NotFound'}
        '409': {$ref: '#/components/responses/Conflict'}
        '422': {$ref: '#/components/responses/ValidationFailed'}
...
components:
...
  securitySchemes:
    cookieSession:
      type: apiKey
      in: cookie
      name: ppr_session
      description: Opaque session token; only its SHA-256 digest is stored.
  responses:
    InvalidCredentials:
      description: Credentials are invalid; account existence and throttle state are intentionally undisclosed.
      content: {application/json: {schema: {$ref: '#/components/schemas/ErrorResponse'}, example: {error: {code: invalid_credentials, message: The credentials are invalid.}}}}
...
    BootstrapDenied:
      description: The bootstrap secret was rejected.
      content: {application/json: {schema: {$ref: '#/components/schemas/ErrorResponse'}}}
    BootstrapUnavailable:
      description: Initial setup is disabled or an account already exists.
      content: {application/json: {schema: {$ref: '#/components/schemas/ErrorResponse'}}}
    ValidationFailed:
      description: An input field failed validation.
      content: {application/json: {schema: {$ref: '#/components/schemas/ErrorResponse'}}}
...
    UnsupportedMediaType:
      description: Content-Type does not match the operation contract (normally application/json or multipart/form-data).
      content: {application/json: {schema: {$ref: '#/components/schemas/ErrorResponse'}}}
...
  schemas:
    BootstrapRequest:
      type: object
      additionalProperties: false
      required: [bootstrap_token, username, email, password]
      properties:
        bootstrap_token: {type: string, minLength: 24, maxLength: 256, writeOnly: true}
        username: {type: string, pattern: '^[a-z0-9][a-z0-9._-]{2,31}$'}
        email: {type: string, format: email, maxLength: 254}
        password: {type: string, minLength: 12, maxLength: 128, writeOnly: true, description: At least 12 Unicode characters and at most 128 UTF-8 bytes.}
    LoginRequest:
      type: object
      additionalProperties: false
      required: [identifier, password]
      properties:
        identifier: {type: string, maxLength: 254}
        password: {type: string, maxLength: 128, writeOnly: true}
...
    UserResponse:
      type: object
      additionalProperties: false
      required: [user]
      properties: {user: {$ref: '#/components/schemas/User'}}
    SessionResponse:
      type: object
      additionalProperties: false
      required: [user, csrf_token, expires_at]
      properties:
        user: {$ref: '#/components/schemas/User'}
        csrf_token: {type: string, description: Session-bound token required on authenticated writes.}
        expires_at: {type: string, format: date-time}
...